Improved UI
Added errors in case of incorrect input from the user in the add items, delete items and also add admin sections. I also added the option to tell the user if the list of the clients is empty or not.
To do: fix the bug when trying to add an admin after multiple logouts