Increased Reusability
Changed the function "iterate", by adding another char paramater.
Before, the function was made to return a string which was starting at the current position and was ending at at the first appeareance of a specified char.
Now, it can be used to return the string which is located between two specified chars.